requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
Flask-SocketIO==5.3.6
fonttools==4.53.1
future==1.0.0
h11==0.14.0
influxdb-client==1.43.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
numpy==2.0.0
numpy-stl==3.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
pyasn1==0.6.0
pygame==2.6.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyserial==3.5
PySimpleGUI==5.0.5
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
reactivex==4.0.4
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
tenacity==9.0.0
tksheet==7.2.12
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
